requirements.txt
Flask==2.3.3
pymongo==4.7.1
flask-cors==4.0.0
python-dotenv==1.0.1
